Create stock_bots_V1
This code is my attempt to get back into programming in Python. I know it is a mess, however, I have not programmed in Python in a while so this is me shaking the rust off. 

In this code, you have two bots talking to one another every 10 seconds for 60 times. The "Seller" will randomly choose an amount to sell a stock price by a percentage-based amount—between -.5% and -5%. The "Seller" will also chose an amount that they want to sell. 

The "Buyer" will always buy either some or all of the shares the seller is selling depending on how much money they have.  This is a rough draft and is not fully completed, but it is a start.